requirements.txt
click
jinja2
anthropic
pyyaml
python-telegram-bot
requests
